Cambios en la velocidad de animación y sonido de pasos.

+++ sprite.h
@@ -11,21 +11,21 @@ class sprite
 		int		   _cantX; // Cantidad de frames en el eje x
 		int		   _cantY; // Cantidad de frames en el eje y
 		Vector2f   _frameSize; //Tamaño del frame.
-		Vector2i   _frameActual; //Frame Actual.
+		Vector2f   _frameActual; //Frame Actual.
 	public:
 		//Constructor
-		sprite(int sprClase, int cantX, int cantY, Vector2i frameActual);
+		sprite(int sprClase, int cantX, int cantY, Vector2f frameActual);
 		//sets
 		void setPosicion(Vector2f position);
 		void setCantX(int cantX);
 		void setCantY(int cantY);		
 		void setFrameSize(Vector2f frameSize);
-		void setFrameActual(Vector2i frameActual);
+		void setFrameActual(Vector2f frameActual);
 		void setFrameX(int frame);
 		void setFrameY(int frame);
 		//gets
 		Sprite getSprite();		
-		Vector2i getFrameActual();
+		Vector2f getFrameActual();
 		Vector2f getFrameSize();
 		Vector2f getPosicion();
 		int getCantX();
